backend/requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
Faker==37.3.0
gunicorn==23.0.0
packaging==25.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.9
